staticx: Pull things out into staticx.elf module
